tests/wpt/web-platform-tests/custom-elements/registering-custom-elements/definition-construction-algorithm-invalid-type.html

<!DOCTYPE html>
<html>
<head>
<title>If TYPE is an invalid custom element type, throw SyntaxError</title>
<meta name="author" title="Aleksei Yu. Semenov" href="mailto:a.semenov@unipro.ru">
<meta name="assert" content="If TYPE is an invalid custom element type, set ERROR to InvalidType and stop.">
<link rel="help" href="http://www.w3.org/TR/custom-elements/#registering-custom-elements">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../testcommon.js"></script>
<link rel="stylesheet" href="/resources/testharness.css">
</head>
<body>
<div id="log"></div>
<script>
test(function() {
    var doc = newHTMLDocument();
    assert_throws('SyntaxError', function() { doc.registerElement('1xa2'); },
                  'Registering invalid custom element type should throw SyntaxError');
}, 'Registering invalid custom element type should throw SyntaxError');
</script>
</body>
</html>
